Remove debian init setting heap dump file
The heap dump file location can be set by the CASSANDRA_HEAPDUMP_DIR
flag via the cassandra-env.sh script. In the debian init script this
value gets shadowed by setting the heap dump file explicitly with the
`-H` flag. This means that there is no clean way, aside from changing
the home directory of the cassandra user, to set the location of the
heap dump file.

This change removes the explicit setting of the heap dump directory via
the `-H` flag and instead moves it into the cassandra-env.sh script. If
the CASSANDRA_HEAPDUMP_DIR flag is not set this should fall back to the
same behavior as before, if it is set then it should respect its value
and set the heap dump directory appropriated.